aws-org/requirements.txt
boto3>=1.26.0
python-dotenv>=1.0.0
Flask>=2.0.0
click>=8.0.0
botocore>=1.29.0
